Finished and tested async PID class
Some of the previous commits in this branch did most of the legwork to get to this point, but they kept throwing a data abort exception. This version fixes that using a static class function that takes a class pointer as an argument. I also discovered that if main exits while there is a sub-process going, it will create a data abort exception.