Updated and tuned the PID
There were a few important things wrong with the PID the way it was. First, it only waited for 1 ms to confirm if it had reached it's target zone. Second, values from the turn PID bled into the drive PID and vice versa. Third, it needed a way to control max speed. Fourth, both versions needed the logger. Fifth, it needed to be tuned to the correct values.