Refactoring and documentation
Our only competition left is Worlds and we have a bit of a break, so tonight we will be having a meeting to discuss what projects we want to work on. In preperation for this, I've done some refactoring and documentation of the flywheel bots code seeing as we will continue using this bot, but I did not do the same for the catapult bot because we will not be using it any more.